\name{seedRNGseq}
\alias{seedRNGseq}
\title{Creating Initial Seed for Random Streams}
\usage{
  seedRNGseq(seed = NULL, normal.kind = NULL,
    verbose = FALSE, version = 2)
}
\arguments{
  \item{seed}{a single or 6-length numeric. If missing then
  the current seed is returned.}

  \item{normal.kind}{Type of Normal random generator. See
  \code{\link{RNG}}.}

  \item{verbose}{logical to toggle verbose messages}
}
\value{
  a 7-length numeric vector.
}
\description{
  This function generate the next random seed used to set
  the first seed in the sequence generated by
  \code{\link{RNGseq}}.
}
\examples{
## generate a seed for RNGseq
# random
seedRNGseq()
seedRNGseq()
seedRNGseq(NULL)
# fixed
seedRNGseq(1)
seedRNGseq(1:6)

# `seedRNGseq(1)` is identical to
set.seed(1)
s <- seedRNGseq()
identical(s, seedRNGseq(1))
\dontshow{ stopifnot(identical(s, seedRNGseq(1))) }
}
\seealso{
  \code{\link{RNGseq}}
}

